# openapi.yaml
openapi: 3.0.0
info:
  title: Calgary Fit Friends
  description: Create pop-up activities and invite family & friends.
  version: "1.3"
servers:
  - url: http://localhost:4201
    description: Local host
paths:
...
  /user:
    get:
      summary: Log a user in
      description: Log a user in.
      parameters:
        - name: name
          in: query
          description: User name
          schema:
            type: string
          required: true
        - name: password
          in: query
          description: User password
          schema:
            type: string
          required: true
      responses:
        "200":
          description: User logged in.
          content:
            text/html:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerMessage"

    post:
      summary: Add a new user.
      description: Add a new user to the system.
      requestBody:
        description: Details of the user to be added.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userInfo:
                  $ref: "#/components/schemas/UserInfo"
                password:
                  type: string
      responses:
        "200":
          description: User successfully added.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userUID:
                    $ref: "#/components/schemas/UID"
        "403":
          description: User already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerMessage"
...
components:
  schemas:
    UserInfo:
      type: object
      properties:
        userUID:
          $ref: "#/components/schemas/UID"
        name:
          type: string
        email:
          type: string
        createdActivities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityInfo"
        joinedActivities:
          type: array
          items:
            $ref: "#/components/schemas/ActivityInfo"
        friendsList:
          type: array
          items:
            $ref: "#/components/schemas/UserInfo"
        reviews:
          type: array
          items:
            type: object
            properties:
              user:
                $ref: "#/components/schemas/UserInfo"
              details:
                type: string
      required:
        - name
    ActivityInfo:
      type: object
      properties:
        activityUID:
          $ref: "#/components/schemas/UID"
        location:
          type: string
        dateTime:
          type: string
        type:
          type: string
        details:
          type: string
      required:
        - location
        - dateTime
        - type
        - details
...
    UID:
      oneOf:
        - type: integer
        - type: string
    ServerMessage:
      type: object
      properties:
        msg:
          type: string
      required:
        - msg